Support Model.upsert()
The Postgres dialect for Sequelize uses a stored procedure to implement
upserts, for compatibility with PostgreSQL before 9.5. CockroachDB has
upsert but no stored procedures, so this monkey patch for the built-in
Sequelize dialect replaces Sequelize's Model.upsert() with a version
that uses `INSERT ... ON CONFLICT DO UPDATE SET`, which CockroachDB does
support.

require()ing the package multiple times should be fine, as `index.js`
is idempotent.

Note that tests will not pass until cockroachdb/cockroach#13962 is
fixed.

Resolves cockroachdb/cockroach#12120